requirements.txt
python-magic~=0.4.27
fastapi~=0.115.12
pillow~=10.4.0
moviepy~=2.1.2
pydantic~=2.11.1
starlette~=0.46.1
pytest~=8.3.5
setuptools~=68.2.0